+++ CPP/src/main.cpp
@@ -7,39 +7,21 @@
 
 using namespace std;
 
+map<int, Card> masterCardList;
+map<string, Profile> profileDB;
+
 int main(){
 
   string tempFileString = "";
-  map<int, Card> masterCardList;
-
+  
   cout << "TEST START" << endl;
+
+  // Load Data
   loadCardData(tempFileString, &masterCardList, true);
-
-  // for (int i = 1001; i < 1007; i++)
-  // {
-  //   cout << "card[1]: ";
-  //   masterCardList.find(i)->second.printCard();
-  //   cout << endl << endl;
-  // }
-
-  map<string, Profile> profileDB;
   loadProfileData(tempFileString, &profileDB, &masterCardList, true);
 
-  Profile tmpProfile = profileDB.at("Funshine");
-  vector<Deck> tmpDeckList = tmpProfile.getAllDecks();
-
-  // for (uint8_t i; i < tmpDeckList.size(); i++)
-  // {
-  //   cout << "Deck: " << tmpDeckList.at(i).getDeckName() << endl;
-
-  //   vector<Card> tmpCardList = tmpDeckList.at(i).getOriginalDeck();
-  //   for (uint8_t j = 0; j < tmpCardList.size(); j++)
-  //   {
-  //     cout << "\t" << (j+1) << "." << tmpCardList.at(j).getCardName();
-  //     cout << endl << endl;
-  //   }
-    
-  // }
+  // Profile tmpProfile = profileDB.at("Funshine");
+  // vector<Deck> tmpDeckList = tmpProfile.getAllDecks();
 
   GameMonitor gm(&masterCardList, &profileDB);
   gm.mainMenu();